Move TableToolbar to seperate file IncidentTableToolbar
